fix(01): extract shared useAppConfig hook and move AppConfig type out of config-debug
